+++ src/main/java/com/soo/routine/repository/BoardRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
 
-    List<Board> findAllByCategoryAndMemberId(String category, int memberId);
+    List<Board> findAllByCategoryAndMemberId(String category, Long memberId);
     List<Board> findAllByCategory(String category);
 
 }
